git-svn-id: svn://svn.code.sf.net/p/x10/code/branches@22656 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10-tools-integration/x10.runtime/src-java/x10/runtime/impl/java/Thread.java
@@ -138,7 +138,7 @@ public class Thread extends java.lang.Thread implements x10.core.RefI {
         throw new UnsupportedOperationException("Cannot serialize " + getClass());
     }
 
-        public int $_get_serialization_id() {
+        public short $_get_serialization_id() {
             throw new UnsupportedOperationException("Cannot serialize " + getClass());
         }
